Change _t_reload to return self, and have establish_relationship_in_target_objects use the reloaded object.
DataMapper's reload method was causing strange behavior in establish_relationship_in_target_objects.  Changing the DataMapper extension to re-fetch the object from the data, and use that instead of the object that was already instantiated, fixed the problem.  Although, I am unsure why this is the case.  I suspect it has something to do with DataMapper's lazy initialization behavior.