Fixed a bug causing t_has_many associates to be wiped out
If the object had not yet loaded its associates (which, since they are
lazy loaded, would not happen unless the associates were used), and
was updated and saved, the associates would be wiped out.  This fix
keeps tracks of which associations have been loaded, and will ensure
that the associates of t_has_many associations have been loaded before
wiping them out and re-saving them.